requirements.txt

Django==3.2.5
requests
psycopg2-binary
Pillow==8.4.0
python-dotenv